Only search getDomSibling if oldStartDom was non-null
This change fixes a nested Portal test case. Only search getDomSibling to find the next startDom after diffing a root node if we have an oldStartDom and it changed parents. oldStartDom can be null meaning future children after the portal should be appended. We should do a getDomSibling in that case.